# docs/swagger.yaml
openapi: 3.0.3
info:
  title: EDI Processing API
  description: |
    A production-ready API for processing EDI (Electronic Data Interchange) files asynchronously.
    
    ## Features
    - Asynchronous EDI file processing
    - Job status tracking
    - Result retrieval
    - Automatic retry logic
    - Health monitoring
    
    ## Workflow
    1. Upload an EDI file via POST /jobs
    2. Receive a job ID
    3. Poll job status via GET /jobs/{job_id}
    4. Retrieve results via GET /jobs/{job_id}/result when completed
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://api.example.com
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Jobs
    description: EDI job processing operations

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check if the API service is healthy and running
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: healthy
                timestamp: '2024-01-15T10:30:00Z'
                service: edi-processing-api

  /jobs:
    post:
      tags:
        - Jobs
      summary: Create a new EDI processing job
      description: |
        Upload an EDI file for asynchronous processing. The file will be validated,
        queued, and processed by a background worker. Returns a job ID for tracking.
      operationId: createJob
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: EDI file to process (max 10MB)
            example:
              file: (binary content)
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateJobResponse'
              example:
                job_id: 65a5b1234567890abcdef123
                message: Job created successfully and queued for processing
        '400':
          description: Bad request - Invalid or missing file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                no_file:
                  value:
                    error: invalid_request
                    message: No file provided or invalid file upload
                empty_file:
                  value:
                    error: invalid_file
                    message: File content is empty
                file_too_large:
                  value:
                    error: file_too_large
                    message: File size exceeds maximum allowed size of 10MB
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                database_error:
                  value:
                    error: database_error
                    message: Failed to save job
                queue_error:
                  value:
                    error: queue_error
                    message: Failed to enqueue job for processing

  /jobs/{job_id}:
    get:
      tags:
        - Jobs
      summary: Get job status
      description: |
        Retrieve the current status of a job by its ID. Status can be:
        - `pending`: Job is queued for processing
        - `processing`: Job is currently being processed
        - `completed`: Job processing completed successfully
        - `failed`: Job processing failed after retries
      operationId: getJobStatus
      parameters:
        - name: job_id
          in: path
          required: true
          description: Unique job identifier (UUID format)
          schema:
            type: string
            format: uuid
          example: 65a5b1234567890abcdef123
      responses:
        '200':
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
              examples:
                pending:
                  value:
                    job_id: 65a5b1234567890abcdef123
                    status: pending
                    retry_count: 0
                    created_at: '2024-01-15T10:30:00Z'
                    updated_at: '2024-01-15T10:30:00Z'
                processing:
                  value:
                    job_id: 65a5b1234567890abcdef123
                    status: processing
                    retry_count: 0
                    created_at: '2024-01-15T10:30:00Z'
                    updated_at: '2024-01-15T10:30:05Z'
                completed:
                  value:
                    job_id: 65a5b1234567890abcdef123
                    status: completed
                    retry_count: 0
                    created_at: '2024-01-15T10:30:00Z'
                    updated_at: '2024-01-15T10:30:10Z'
                    result:
                      claims:
                        - claim_id: CLM001
                          member_id: MEM123
                          amount: 2500
                      summary:
                        total_claims: 3
                        total_amount: 7000
                failed:
                  value:
                    job_id: 65a5b1234567890abcdef123
                    status: failed
                    error: "Parsing failed: invalid EDI format"
                    retry_count: 3
                    created_at: '2024-01-15T10:30:00Z'
                    updated_at: '2024-01-15T10:30:20Z'
        '400':
          description: Bad request - Invalid job ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: invalid_job_id
                message: Invalid job ID format
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: job_not_found
                message: Job with ID 65a5b1234567890abcdef123 not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /jobs/{job_id}/result:
    get:
      tags:
        - Jobs
      summary: Get job result
      description: |
        Retrieve the processing result for a completed job. If the job is not yet
        completed, returns the current status. Only completed jobs have results available.
      operationId: getJobResult
      parameters:
        - name: job_id
          in: path
          required: true
          description: Unique job identifier (UUID format)
          schema:
            type: string
            format: uuid
          example: 65a5b1234567890abcdef123
      responses:
        '200':
          description: Job result retrieved successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ResultResponse'
                  - $ref: '#/components/schemas/PendingResultResponse'
              examples:
                completed:
                  value:
                    status: completed
                    claims:
                      - claim_id: CLM001
                        member_id: MEM123
                        amount: 2500
                      - claim_id: CLM002
                        member_id: MEM456
                        amount: 3000
                      - claim_id: CLM003
                        member_id: MEM789
                        amount: 1500
                    summary:
                      total_claims: 3
                      total_amount: 7000
                pending:
                  value:
                    job_id: 65a5b1234567890abcdef123
                    status: pending
                    message: Job is pending. Result not available yet.
                processing:
                  value:
                    job_id: 65a5b1234567890abcdef123
                    status: processing
                    message: Job is processing. Result not available yet.
        '400':
          description: Bad request - Invalid job ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: invalid_job_id
                message: Invalid job ID format
        '404':
          description: Job or result not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                job_not_found:
                  value:
                    error: job_not_found
                    message: Job with ID 65a5b1234567890abcdef123 not found
                result_not_found:
                  value:
                    error: result_not_found
                    message: Result for job 65a5b1234567890abcdef123 not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - timestamp
        - service
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          description: Health status of the service
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the health check
        service:
          type: string
          description: Service name

    CreateJobResponse:
      type: object
      required:
        - job_id
        - message
      properties:
        job_id:
          type: string
          description: Unique identifier for the created job
          example: 65a5b1234567890abcdef123
        message:
          type: string
          description: Success message
          example: Job created successfully and queued for processing

    JobStatusResponse:
      type: object
      required:
        - job_id
        - status
        - retry_count
        - created_at
        - updated_at
      properties:
        job_id:
          type: string
          description: Unique job identifier
        status:
          type: string
          enum: [pending, processing, completed, failed]
          description: Current status of the job
        result:
          $ref: '#/components/schemas/Result'
          description: Processing result (only present when status is completed)
        error:
          type: string
          description: Error message (only present when status is failed)
        retry_count:
          type: integer
          description: Number of retry attempts
          example: 0
        created_at:
          type: string
          format: date-time
          description: Job creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp

    ResultResponse:
      type: object
      required:
        - status
        - claims
        - summary
      properties:
        status:
          type: string
          enum: [completed]
          description: Result status
        claims:
          type: array
          description: Array of parsed claims
          items:
            $ref: '#/components/schemas/Claim'
        summary:
          $ref: '#/components/schemas/Summary'

    PendingResultResponse:
      type: object
      required:
        - job_id
        - status
        - message
      properties:
        job_id:
          type: string
          description: Job identifier
        status:
          type: string
          enum: [pending, processing, failed]
          description: Current job status
        message:
          type: string
          description: Status message

    Result:
      type: object
      properties:
        claims:
          type: array
          items:
            $ref: '#/components/schemas/Claim'
        summary:
          $ref: '#/components/schemas/Summary'

    Claim:
      type: object
      required:
        - claim_id
        - member_id
        - amount
      properties:
        claim_id:
          type: string
          description: Unique claim identifier
          example: CLM001
        member_id:
          type: string
          description: Member identifier
          example: MEM123
        amount:
          type: number
          format: double
          description: Claim amount
          example: 2500.00

    Summary:
      type: object
      required:
        - total_claims
        - total_amount
      properties:
        total_claims:
          type: integer
          description: Total number of claims
          example: 3
        total_amount:
          type: number
          format: double
          description: Total amount of all claims
          example: 7000.00

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: invalid_request
        message:
          type: string
          description: Human-readable error message
          example: No file provided or invalid file upload